Create plagiarismChecker
The skeleton plagiarism checker python script. The main reason behind this check to be done as a python script as against adding it as a rule in LangaugeTool is purely due to faster deployment and development. Later on the same checker will be made available as a LanguageTool api and then used in the python wrapper, language-check.